Strip everything but ena
