--- app/(datasets)/stories/hub.tsx
@@ -1,21 +1,0 @@
-'use client';
-import React from 'react';
-import { StoriesHubContent, useFiltersWithQS } from '@lib';
-import Providers from '../providers';
-
-export default function Hub({ stories: allStories }: { stories: any }) {
-  const controlVars = useFiltersWithQS();
-
-  return (
-    <Providers>
-      <StoriesHubContent
-        allStories={allStories}
-        onFilterchanges={() => controlVars}
-        storiesString={{
-          one: 'story',
-          other: 'stories',
-        }}
-      />
-    </Providers>
-  );
-}
